GPX-693: refactor implemented new way to collect rbac data from cluster.
In the old approach the collector sent for every namespace and subject a respective SubjectAccessReview which lead to rate-limiting by the kubernetes api.

In the new approach the collector grabs all cluster/-role/-bindings and resolves the permission locally. This lead to a massive performance increase and uses way less resources.