GPX-693: refactor Added server
Entrypoint of the serve command.
Starts up a new Server.
Lists all resources, continuously watches them via the k8s watch api.
Handles the incoming signal from the resources to updated the configmap with the new permissions if there are any changes to the previous configmap.

Also drains the signal channel as rbac data gets added in bursts. Leaves one in the channel so collection gets triggered one more time. Reasons for that is if there is a new signal coming in while draining we want to act on it. Thus, waits 5 seconds.